Changes to be committed: 	deleted:    api/src/database/seeds/directus_directus_fields_2025-04-09_205004.sql 	modified:   api/src/services/documents.ts 	modified:   api/src/services/rags.ts 	modified:   api/src/services/servers.ts
